Split BaseRunner test out into own file, for maintainability
